Add source .proto file for update_metadata_pb2.py
*_pb2.py files are autogenerated by the protobuf compiler.  Sometimes
the format changes in a way that is incompatible with the
python-protobuf runtime, so it is unwise to maintain *just* the
compiled version in our tree.  Add the corresponding source file and
a Makefile (borrowed from carriersettings-extractor) to let the user
easily regenerate it from source.

I found update_metadata.proto in this repo:
https://android.googlesource.com/platform/system/update_engine

By comparing the compiled output I determined that our current version
of update_metadata_pb2.py contains the deprecation flags introduced in
upstream commit 0f59a9a41177, and it does not contain the
unpadded_signature_size field added in the subsequent commit
(7bbe015a1bd1c).  So I synced to 0f59a9a41177.

Fixes AOSPAlliance/android-prepare-vendor#58.